+++ src/Autodesk.RefineryToolkits.sln
@@ -11,9 +11,9 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SpacePlanningTests", "..\te
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Core", "Core\Core.csproj", "{A528FCD0-0855-409C-88E0-E36E6A718FD3}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MassingToolkitRevit", "Massing\MassingToolkitRevit.csproj", "{290CA621-CA65-4A16-8503-B9E0C63EB739}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MassingRevit", "MassingRevit\MassingRevit.csproj", "{290CA621-CA65-4A16-8503-B9E0C63EB739}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MassingToolkitSandbox", "MassingSandbox\MassingToolkitSandbox.csproj", "{89DF56A1-A27B-43D9-9D70-B650B2DB4B64}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MassingSandbox", "MassingSandbox\MassingSandbox.csproj", "{89DF56A1-A27B-43D9-9D70-B650B2DB4B64}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -87,16 +87,16 @@ Global
 		{290CA621-CA65-4A16-8503-B9E0C63EB739}.Release|x64.Build.0 = Release|x64
 		{89DF56A1-A27B-43D9-9D70-B650B2DB4B64}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{89DF56A1-A27B-43D9-9D70-B650B2DB4B64}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{89DF56A1-A27B-43D9-9D70-B650B2DB4B64}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{89DF56A1-A27B-43D9-9D70-B650B2DB4B64}.Debug|x64.Build.0 = Debug|Any CPU
+		{89DF56A1-A27B-43D9-9D70-B650B2DB4B64}.Debug|x64.ActiveCfg = Debug|x64
+		{89DF56A1-A27B-43D9-9D70-B650B2DB4B64}.Debug|x64.Build.0 = Debug|x64
 		{89DF56A1-A27B-43D9-9D70-B650B2DB4B64}.DebugRevit|Any CPU.ActiveCfg = Debug|Any CPU
 		{89DF56A1-A27B-43D9-9D70-B650B2DB4B64}.DebugRevit|Any CPU.Build.0 = Debug|Any CPU
-		{89DF56A1-A27B-43D9-9D70-B650B2DB4B64}.DebugRevit|x64.ActiveCfg = Debug|Any CPU
-		{89DF56A1-A27B-43D9-9D70-B650B2DB4B64}.DebugRevit|x64.Build.0 = Debug|Any CPU
+		{89DF56A1-A27B-43D9-9D70-B650B2DB4B64}.DebugRevit|x64.ActiveCfg = Debug|x64
+		{89DF56A1-A27B-43D9-9D70-B650B2DB4B64}.DebugRevit|x64.Build.0 = Debug|x64
 		{89DF56A1-A27B-43D9-9D70-B650B2DB4B64}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{89DF56A1-A27B-43D9-9D70-B650B2DB4B64}.Release|Any CPU.Build.0 = Release|Any CPU
-		{89DF56A1-A27B-43D9-9D70-B650B2DB4B64}.Release|x64.ActiveCfg = Release|Any CPU
-		{89DF56A1-A27B-43D9-9D70-B650B2DB4B64}.Release|x64.Build.0 = Release|Any CPU
+		{89DF56A1-A27B-43D9-9D70-B650B2DB4B64}.Release|x64.ActiveCfg = Release|x64
+		{89DF56A1-A27B-43D9-9D70-B650B2DB4B64}.Release|x64.Build.0 = Release|x64
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
