+++ src/MassingRevit/MassingRevit.ruleset
@@ -1,5 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
-<RuleSet Name="Rules for MassingToolkit" Description="Code analysis rules for MassingToolkit.csproj." ToolsVersion="15.0">
+<RuleSet Name="Rules for GenerativeToolkit" Description="Code analysis rules for GenerativeToolkit.csproj." ToolsVersion="15.0">
   <Rules AnalyzerId="Microsoft.Analyzers.ManagedCodeAnalysis" RuleNamespace="Microsoft.Rules.Managed">
     <Rule Id="CA1000" Action="Warning" />
     <Rule Id="CA1001" Action="Warning" />
