+++ src/MassingRevit/Properties/AssemblyInfo.cs
