+++ src/MassingRevit/pkg.json
@@ -1,9 +1,9 @@
 {
   "license": "MIT",
   "file_hash": null,
-  "name": "Refinery Nodes",
+  "name": "Refinery Toolkits - MassingRevit",
   "version": "2019.3.1",
-  "description": "Refinery nodes.",
+  "description": "Generative Design",
   "group": "Proving Ground",
   "keywords": [
     "utilities"
@@ -19,7 +19,7 @@
   "repository_url": "",
   "contains_binaries": true,
   "node_libraries": [
-    "GenerativeToolkitRevit, Version=2019.3.1.0, Culture=neutral, PublicKeyToken=null",
-    "GenerativeToolkitSandbox, Version=2019.3.1.0, Culture=neutral, PublicKeyToken=null"
+    "Autodesk.RefineryToolkits.MassingRevit, Version=2019.3.1.0, Culture=neutral, PublicKeyToken=null",
+    "Autodesk.RefineryToolkits.MassingSandbox, Version=2019.3.1.0, Culture=neutral, PublicKeyToken=null"
   ]
 }
