--- src/MassingSandbox/MassingToolkitSandbox_DynamoCustomization.xml
@@ -1,11 +1,0 @@
-﻿<?xml version="1.0"?>
-<doc>
-  <assembly>
-    <name>MassingToolkitSandbox</name>
-  </assembly>
-  <namespaces>
-    <namespace name="MassingToolkit">
-      <category>Massing Toolkit</category>
-    </namespace>
-  </namespaces>
-</doc>
