+++ src/MassingSandbox/packages.config
@@ -2,6 +2,7 @@
 <packages>
   <package id="CommonServiceLocator" version="1.0" targetFramework="net472" />
   <package id="DynamoVisualProgramming.Core" version="2.2.0.4565" targetFramework="net472" />
+  <package id="DynamoVisualProgramming.DynamoCoreNodes" version="2.2.0.4565" targetFramework="net472" />
   <package id="DynamoVisualProgramming.DynamoServices" version="2.2.0.4565" targetFramework="net472" />
   <package id="DynamoVisualProgramming.ZeroTouchLibrary" version="2.2.0.4565" targetFramework="net472" />
   <package id="Microsoft.CodeAnalysis.FxCopAnalyzers" version="2.6.3" targetFramework="net45" developmentDependency="true" />
